refactor: extract sessionruntime idle-unload scheduling
Move the runtime idle-unload timer type, delays, and scheduling/cancellation
logic out of service.go into idle_unload.go.
