+++ server/sessionruntime/service.go
@@ -51,16 +51,6 @@ type Service struct {
 	idleTimers             map[string]*runtimeIdleTimer
 }
 
-type runtimeIdleTimer struct {
-	generation uint64
-	timer      *time.Timer
-}
-
-const (
-	defaultRuntimeIdleUnloadDelay     = 5 * time.Second
-	defaultRunFinishedIdleUnloadDelay = 3 * time.Minute
-)
-
 func NewService(persistenceRoot string, metadataStore *metadata.Store, authManager *auth.Manager, fastModeState *runtime.FastModeState, background *shelltool.Manager, backgroundRouter *runtimewire.BackgroundEventRouter, runtimes *registry.RuntimeRegistry, sessionStores *registry.SessionStoreRegistry, storeOptions ...session.StoreOption) *Service {
 	svc := &Service{
 		persistenceRoot:        strings.TrimSpace(persistenceRoot),
@@ -541,132 +531,6 @@ func (s *Service) markClaimOrphaned(sessionID string, claim *registry.RuntimeCla
 	}
 	claim.DropOwner(ownerID)
 	s.scheduleIdleUnload(strings.TrimSpace(sessionID), s.defaultIdleUnloadDelay())
-}
-
-func (s *Service) runtimeInterestChanged(sessionID string, reason registry.RuntimeInterestReason) {
-	delay := s.defaultIdleUnloadDelay()
-	if reason == registry.RuntimeInterestRunFinished {
-		delay = s.runFinishedIdleUnloadDelay()
-	}
-	s.scheduleIdleUnload(sessionID, delay)
-}
-
-func (s *Service) cancelScheduledIdleUnload(sessionID string) {
-	if s == nil {
-		return
-	}
-	trimmedSessionID := strings.TrimSpace(sessionID)
-	if trimmedSessionID == "" {
-		return
-	}
-	s.mu.Lock()
-	state := s.idleTimers[trimmedSessionID]
-	if state != nil {
-		state.generation++
-		if state.timer != nil {
-			state.timer.Stop()
-		}
-	}
-	s.mu.Unlock()
-}
-
-func (s *Service) clearScheduledIdleUnload(sessionID string) {
-	if s == nil {
-		return
-	}
-	trimmedSessionID := strings.TrimSpace(sessionID)
-	if trimmedSessionID == "" {
-		return
-	}
-	s.mu.Lock()
-	state := s.idleTimers[trimmedSessionID]
-	if state != nil && state.timer != nil {
-		state.timer.Stop()
-	}
-	delete(s.idleTimers, trimmedSessionID)
-	s.mu.Unlock()
-}
-
-func (s *Service) defaultIdleUnloadDelay() time.Duration {
-	if s == nil || s.idleUnloadDelay <= 0 {
-		return defaultRuntimeIdleUnloadDelay
-	}
-	return s.idleUnloadDelay
-}
-
-func (s *Service) runFinishedIdleUnloadDelay() time.Duration {
-	if s == nil || s.runFinishedUnloadDelay <= 0 {
-		return defaultRunFinishedIdleUnloadDelay
-	}
-	return s.runFinishedUnloadDelay
-}
-
-func (s *Service) scheduleIdleUnload(sessionID string, delay time.Duration) {
-	if s == nil {
-		return
-	}
-	trimmedSessionID := strings.TrimSpace(sessionID)
-	if trimmedSessionID == "" || delay <= 0 {
-		return
-	}
-	s.mu.Lock()
-	if s.idleTimers == nil {
-		s.idleTimers = make(map[string]*runtimeIdleTimer)
-	}
-	state := s.idleTimers[trimmedSessionID]
-	if state == nil {
-		state = &runtimeIdleTimer{}
-		s.idleTimers[trimmedSessionID] = state
-	}
-	state.generation++
-	generation := state.generation
-	if state.timer != nil {
-		state.timer.Stop()
-	}
-	state.timer = time.AfterFunc(delay, func() {
-		s.runScheduledIdleUnload(trimmedSessionID, generation)
-	})
-	s.mu.Unlock()
-}
-
-func (s *Service) runScheduledIdleUnload(sessionID string, generation uint64) {
-	if s == nil {
-		return
-	}
-	trimmedSessionID := strings.TrimSpace(sessionID)
-	if trimmedSessionID == "" {
-		return
-	}
-	s.mu.Lock()
-	state := s.idleTimers[trimmedSessionID]
-	if state == nil || state.generation != generation {
-		s.mu.Unlock()
-		return
-	}
-	s.mu.Unlock()
-	claim := s.runtimes.RuntimeClaimFor(trimmedSessionID)
-	if claim == nil || claim.Closing() || claim.OwnerCount() > 0 {
-		return
-	}
-	if s.runtimeHasSubscribers(trimmedSessionID) {
-		return
-	}
-	if active, err := s.runtimeHasActiveRun(context.Background(), trimmedSessionID); err != nil || active {
-		return
-	}
-	_, _ = s.ReleaseSessionRuntime(context.Background(), serverapi.SessionRuntimeReleaseRequest{
-		ClientRequestID: uuid.NewString(),
-		SessionID:       trimmedSessionID,
-		OnlyIfIdle:      true,
-		DropOwner:       true,
-	})
-}
-
-func (s *Service) runtimeHasSubscribers(sessionID string) bool {
-	if s == nil || s.runtimes == nil {
-		return false
-	}
-	return s.runtimes.HasRuntimeSubscribers(strings.TrimSpace(sessionID))
 }
 
 func (s *Service) HasActiveRun(ctx context.Context, sessionID string) (bool, error) {
